Move to StringExtensions
